dashboard/requirements.txt
Flask==3.0.2
gunicorn==21.2.0
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.2.2
joblib==1.4.2
streamlit==1.41.1
